pyproject.toml
[project]
name = "llm-council"
version = "0.1.0"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "requests>=2.32.5",
    "clerk-backend-api>=4.0.0",
    "stripe>=14.0.1",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
    "asyncpg>=0.31.0",
    "sqlalchemy>=2.0.44",
    "psycopg2-binary>=2.9.11",
]
